Add files via upload
This PPT was used for one of my business analysis presentations in one of the companies that I have applied as part of their recruitment process. I was tasked to create a documentation for the Visitor Management Systems that aims to replace the traditional 'logbook' used by the security team. After 3 days of working, I was able to come up with a specification document and test cases (which I lack experience with). 

The design is heavily inspired by one of Sir Jordan Deja's work. You can check his Github repository via this link https://jrdndj.github.io/.

Thanky you!